add expect_download
